pyproject.toml
[tool.poetry]
name = "shop-app"
version = "0.1.0"
description = ""
authors = ["gazin66 <gazin9875@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.6"
fastapi = "^0.112.2"
sqlalchemy = "^2.0.34"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
httpx = "^0.27.2"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
